requirements.txt
aiohttp==3.6.3
async-timeout==3.0.1
attrs==20.3.0
autopep8==1.5.4
beautifulsoup4==4.9.3
certifi==2020.12.5
chardet==3.0.4
discord.py==1.5.1
idna==2.10
multidict==4.7.6
Pillow==8.0.1
pip==20.3.1
pycodestyle==2.6.0
requests==2.25.0
setuptools==49.2.1
soupsieve==2.0.1
toml==0.10.2
urllib3==1.26.2
wheel==0.36.1
yarl==1.5.1